refactor: split geo_index/mod.rs into separate files
The geo_index module was a single 2134-line file. Split it into focused
files for better readability and maintainability:

- mod.rs: GeoMapIndex enum definition and core methods (~290 lines)
- builders.rs: GeoMapIndexMmapBuilder and GeoMapIndexGridstoreBuilder
- payload_index.rs: ValueIndexer, PayloadFieldIndex, PayloadFieldIndexRead impls
- tests.rs: all test code (~1340 lines)

No logic changes; all 40 existing tests pass.
